Mgmtworker: merge ClusterService into Service handler
It turns out that every Service task needs to be broadcast to all
mgmtworkers anyway, so there's no point in having one separate
for the cluster (who's original point was the broadcasting).

Put the cluster-related tasks into the regular service task handler.

So there's also no need for a separate "cluster queue" name, we just
use a service queue name per MANAGER_NAME.

(also dropping --name because that is never passed in anyway)
